Implemented Pull-up Field refactoring opportunity detector
Implemented Pull-up field detector using a custom listener.
Everything is ad hoc so you probably need to check for duplicate codes and change the code a bit if you want to merge this and reuse the old listeners.
Also created a main file which is basically a "how to use this" script.
Added the requested package name and file path values to the output.json file.
Code is tested and works on my machine. :grinning: